Use +groupGoName for informers interface pkgName
If a group name includes a "-", then the previous implementation created
invalid package names. This allows users to provide a better/valid name.

client-gen uses the groupName commentTag to handle the same thing.
informer-gen already uses the groupGoName comment tag for a few other
things, so this commit makes it use that for internal and external
packages as well.